python-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdf2docx==0.5.8
pdfplumber==0.11.4
openpyxl==3.1.5
python-pptx==1.0.2
PyMuPDF==1.25.3
ebooklib==0.18
mammoth==1.8.0
